// src/platform/packages/shared/kbn-unified-data-table/src/types.d.ts
import type { FunctionComponent } from 'react';
import type { EuiContextMenuItem, EuiDataGridCellValueElementProps, EuiDataGridColumn } from '@elastic/eui';
import type { DataTableRecord, DataTableColumnsMeta } from '@kbn/discover-utils/src/types';
import type { DataView } from '@kbn/data-views-plugin/common';
import type { FieldFormatsStart } from '@kbn/field-formats-plugin/public';
export type { DataTableColumnsMeta } from '@kbn/discover-utils/types';
export type { DataGridDensity } from './constants';
/**
 * User configurable state of data grid, persisted in saved search
 */
export interface UnifiedDataTableSettings {
    columns?: Record<string, UnifiedDataTableSettingsColumn>;
}
export interface UnifiedDataTableSettingsColumn {
    width?: number;
    /**
    Optional props passed to Columns to display provided labels as column names instead of field names.
    This object maps column field names to their corresponding display labels.
    These labels will take precedence over the data view field names.
    */
    display?: string;
}
export type ValueToStringConverter = (rowIndex: number, columnId: string, options?: {
    compatibleWithCSV?: boolean;
}) => {
    formattedString: string;
    withFormula: boolean;
};
export type DataGridCellValueElementProps = EuiDataGridCellValueElementProps & {
    row: DataTableRecord;
    dataView: DataView;
    fieldFormats: FieldFormatsStart;
    closePopover: () => void;
    isCompressed?: boolean;
    columnsMeta: DataTableColumnsMeta | undefined;
};
export type CustomCellRenderer = Record<string, FunctionComponent<DataGridCellValueElementProps>>;
export interface CustomGridColumnProps {
    column: EuiDataGridColumn;
    headerRowHeight?: number;
}
export type CustomGridColumnsConfiguration = Record<string, (props: CustomGridColumnProps) => EuiDataGridColumn>;
export type DataGridPaginationMode = 'multiPage' | 'singlePage' | 'infinite';
export type CustomBulkActions = Array<Omit<React.ComponentProps<typeof EuiContextMenuItem>, 'onClick'> & {
    onClick: (payload: {
        selectedDocIds: string[];
    }) => void;
    label: React.ReactElement | string;
    key: string;
}>;